tetragon: Remove unused cgo code for reading perf events
We switched to cilium/ebpf for reading perf ring buffer some
time ago [1], but we did not remove the old ring buffer code,
removing it now.

Keeping just PerfEventConfig type used by observer, attr objects
used by map_linux.go and few helper functions used by other code.

[1] 6dd4ee1022eb ("pkg/observer: remove old runEvents code")
