+++ pkg/bpf/map_linux.go
@@ -33,6 +33,20 @@ import (
 	"golang.org/x/sys/unix"
 )
 
+type bpfAttrObjOp struct {
+	pathname uint64
+	fd       uint32
+	pad0     [4]byte
+}
+
+type bpfAttrMapOpElem struct {
+	mapFd uint32
+	pad0  [4]byte
+	key   uint64
+	value uint64 // union: value or next_key
+	flags uint64
+}
+
 type MapKey interface {
 	fmt.Stringer
